+++ packages/docs/src/components/footer.js
@@ -1,18 +1,18 @@
 /** @jsx jsx */
-import { jsx, Footer, Container, Flex } from 'theme-ui'
+import { jsx, Footer, Container } from 'theme-ui'
 import NavLink from './nav-link'
 
 export default props => (
   <Footer
     sx={{
-      pt: 4,
-    }}
-  >
+      py: 3,
+    }}>
     <Container>
-      <Flex>
+      <div sx={{ display: 'flex' }}>
+        <div sx={{ mx: 'auto' }} />
         <NavLink to="/">Theme UI</NavLink>
         <NavLink href="https://github.com/system-ui/theme-ui">GitHub</NavLink>
-      </Flex>
+      </div>
     </Container>
   </Footer>
 )