+++ packages/gatsby-plugin-theme-ui/gatsby-ssr.js
@@ -8,8 +8,7 @@ const noflash = `
     var mode = localStorage.getItem('theme-ui-color-mode');
     if (!mode) return
     document.body.classList.add('theme-ui-' + mode);
-  } catch (e) {
-  }
+  } catch (e) {}
 })();
 `