+++ packages/theme-ui/src/color-modes.js
@@ -23,9 +23,9 @@ export const useColorState = initialMode => {
   useEffect(() => {
     // initialize
     const stored = storage.get()
-    const dark = getMediaQuery()
-    if (dark) setMode('dark')
     document.body.classList.remove('theme-ui-' + stored)
+    const dark = getMediaQuery()
+    if (!stored && dark) return setMode('dark')
     if (!stored || stored === mode) return
     setMode(stored)
   }, [])
@@ -54,10 +54,9 @@ const bodyColor = (theme = {}) => {
   const styles = {}
   Object.keys(modes).forEach(mode => {
     const colors = modes[mode]
-    styles[`&.theme-ui-${mode}`] = {
-      color: colors.text,
-      bg: colors.background,
-    }
+    styles[`&.theme-ui-${mode}`] = theme.useCustomProperties
+      ? colors
+      : { color: colors.text, bg: colors.background }
   })
 
   return css({