+++ packages/theme-ui/src/provider.js
@@ -10,6 +10,39 @@ import { createComponents } from './components'
 
 const mergeState = (state, next) => merge.all({}, state, next)
 
+const colorModesToCSSProperties = modes => {
+  return Object.keys(modes).reduce((parsedModes, modeKey) => {
+    const colors = modes[modeKey]
+    return {
+      ...parsedModes,
+      [modeKey]: Object.keys(colors).reduce(
+        (parsedColors, colorKey) => ({
+          ...parsedColors,
+          [`--theme-ui-${colorKey}`]: colors[colorKey],
+        }),
+        {}
+      ),
+    }
+  }, {})
+}
+
+const applyCSSProperties = theme => {
+  const { colors } = theme
+  return {
+    ...theme,
+    colors: Object.keys(colors).reduce(
+      (parsedColors, key) => ({
+        ...parsedColors,
+        [key]:
+          key === 'modes'
+            ? colorModesToCSSProperties(colors[key])
+            : `var(--theme-ui-${key}, ${colors[key]})`,
+      }),
+      {}
+    ),
+  }
+}
+
 const applyColorMode = (theme, mode) => {
   if (!mode) return theme
   const modes = get(theme, 'colors.modes', {})
@@ -22,7 +55,7 @@ const BaseProvider = ({ context, components, children }) => {
   const theme = applyColorMode(context.theme, context.colorMode)
   return jsx(
     EmotionContext.Provider,
-    { value: theme },
+    { value: theme.useCustomProperties ? applyCSSProperties(theme) : theme },
     jsx(
       MDXProvider,
       { components },