+++ packages/theme-ui/test/color-modes.js
@@ -32,8 +32,7 @@ test('renders with color modes', () => {
       <ThemeProvider
         theme={{
           initialColorMode: 'light',
-        }}
-      >
+        }}>
         <Mode />
       </ThemeProvider>
     )
@@ -93,8 +92,7 @@ test('color mode is passed through theme context', () => {
             },
           },
         },
-      }}
-    >
+      }}>
       <Button />
     </ThemeProvider>
   )
@@ -102,6 +100,37 @@ test('color mode is passed through theme context', () => {
   button.click()
   expect(mode).toBe('dark')
   expect(tree.getByText('test')).toHaveStyleRule('color', 'cyan')
+})
+
+test('converts color modes to css properties', () => {
+  const Box = props => (
+    <div
+      sx={{
+        color: 'text',
+      }}
+      children="test"
+    />
+  )
+  const tree = render(
+    <ThemeProvider
+      theme={{
+        useCustomProperties: true,
+        colors: {
+          text: '#000',
+          modes: {
+            dark: {
+              text: '#fff',
+            },
+          },
+        },
+      }}>
+      <Box />
+    </ThemeProvider>
+  )
+  expect(tree.getByText('test')).toHaveStyleRule(
+    'color',
+    'var(--theme-ui-text,#000)'
+  )
 })
 
 test('does not initialize mode', () => {
@@ -146,13 +175,11 @@ test('inherits color mode state from parent context', () => {
     <ThemeProvider
       theme={{
         initialColorMode: 'outer',
-      }}
-    >
+      }}>
       <ThemeProvider
         theme={{
           initialColorMode: 'inner',
-        }}
-      >
+        }}>
         <Consumer />
       </ThemeProvider>
     </ThemeProvider>
@@ -244,8 +271,7 @@ test('ColorMode component renders with colors', () => {
             },
           },
         },
-      }}
-    >
+      }}>
       <ColorMode />
     </ThemeProvider>
   )