+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## Unreleased
+
+## v0.2.10 2019-07-08
+
+- Fix keys in tailwind preset #171
+
+## v0.2.9 2019-07-08
+
+- Add optional support for CSS custom properties #166
+
+## v0.2.8 2019-07-06
+
+- `@theme-ui/sidenav` initial publish
+- `@theme-ui/prism` add `display: inline-block` to keep empty lines
+
+## v0.2.7 2019-07-05
+
+- `@theme-ui/prism` pass outer className to element #163
+
+## v0.2.6 2019-07-04
+
+- Adjust color mode initialization from media query #157
+
+## v0.2.5 2019-07-03
+
+- Fix publish
+
+## v0.2.4 2019-07-03
+
+- Adjust microbundle setup for @theme-ui/prism
+
+## v0.2.3 2019-07-02
+
+- Add @theme-ui/prism package
+
 ## v0.2.2 2019-07-02
 
 - Add `key` prop to element in gatsby-plugin-theme-ui #145